% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_DEP.R
\name{add_rejections_no_p.adj}
\alias{add_rejections_no_p.adj}
\title{add_rejections_no_p.adj}
\usage{
add_rejections_no_p.adj(data.se, alpha = 0.05, lfc = 1)
}
\arguments{
\item{lfc}{}
}
\description{
reference to DEP::add_rejections(),which not support significant with out p adjust,
this function as supplymentary
}
