% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_Chemmine.R
\name{add_sdf_igraph_highlight}
\alias{add_sdf_igraph_highlight}
\title{add_sdf_igraph_highlight}
\usage{
add_sdf_igraph_highlight(sdf.igraph, highlight = NULL)
}
\arguments{
\item{sdf.igraph}{igraph}

\item{highlight}{vector}
}
\value{
igraph
}
\description{
add_sdf_igraph_highlight
}
