% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatisticFunction.R
\name{analyzePathwayGlobalTest}
\alias{analyzePathwayGlobalTest}
\title{analyzePathwayGlobalTest}
\usage{
analyzePathwayGlobalTest(pathway.matrix, pathway.group)
}
\arguments{
\item{pathway.matrix}{should be a matrix, sample as rowname, kegg id as colname}

\item{pathway.group}{should be a vector with length same as nrow(pathway.matrix), indicate group of sample}
}
\value{
global.test.result
}
\description{
pathway enrichment by global test
}
