% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatisticFunction.R
\name{analyzePathwayHypertest}
\alias{analyzePathwayHypertest}
\title{analyzePathwayHypertest}
\usage{
analyzePathwayHypertest(kegg.id)
}
\arguments{
\item{kegg.id}{}
}
\description{
analyzePathwayHypertest
}
