% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_CFM.R
\name{CFM_data_get_atom_map}
\alias{CFM_data_get_atom_map}
\title{Atom tracing map}
\usage{
CFM_data_get_atom_map(object, iso_ele = "[13]C", BPPARAM = SerialParam())
}
\arguments{
\item{object}{cfmd}

\item{iso_ele}{'\link{13}C'}

\item{BPPARAM}{BiocParallel::SerialParam()}
}
\value{
cfmd
}
\description{
Atom tracing map
}
\section{Functions}{
\itemize{
\item \code{CFM_data_get_atom_map()}: get atom map

}}
\note{
CFM_data_get_atom_map take too much times when construct, major in get_CFM_data_trans_map>>>fmcs
}
