% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_inhouse_database_single_injection.R
\name{build_inhouse_database_single_injection}
\alias{build_inhouse_database_single_injection}
\title{build inhouse database}
\usage{
build_inhouse_database_single_injection(compound.record.file)
}
\arguments{
\item{compound.record.file}{paht of xlsx file, contain column pubchem.cid}
}
\description{
Detect ion peaks from compounds standard. This method just read in single injection (pos and neg) and detect possible adduct form mz.
Usually can not determine whether peaks generated from compounds.
}
