% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_CFM.R
\name{CFM_predict}
\alias{CFM_predict}
\alias{CFM_annotate}
\alias{CFM_annotate_by_fraggen}
\alias{CFM_annotate_by_predict}
\alias{CFM_fraggen}
\alias{read_CFM_predict_result}
\title{CFM predict}
\usage{
CFM_predict(
  smiles_or_inchi_or_file = "[H]C1(O)O[C@]([H])(CO)[C@@]([H])(O)[C@]([H])(O)[C@@]1([H])O",
  prob_thresh = 0.001,
  param_adduct = "[M+H]+",
  annotate_fragments = 1,
  output_file_or_dir = NULL,
  apply_postproc = 0,
  suppress_exceptions = 1
)

CFM_annotate(
  smiles_or_inchi = "[H]C1(O)O[C@]([H])(CO)[C@@]([H])(O)[C@]([H])(O)[C@@]1([H])O",
  spectrum_file = NULL,
  id = "AN_ID",
  ppm_mass_tol = 5,
  abs_mass_tol = 0,
  param_adduct = "[M+H]+",
  output_file = NULL,
  ...
)

CFM_annotate_by_fraggen(
  smiles_or_inchi = "[H]C1(O)O[C@]([H])(CO)[C@@]([H])(O)[C@]([H])(O)[C@@]1([H])O",
  spectrum_file = NULL,
  max_depth = 1,
  id = "AN_ID",
  ppm_mass_tol = 5,
  abs_mass_tol = 0,
  param_adduct = "[M+H]+",
  output_file = NULL,
  ...
)

CFM_annotate_by_predict(
  smiles_or_inchi = "[H]C1(O)O[C@]([H])(CO)[C@@]([H])(O)[C@]([H])(O)[C@@]1([H])O",
  id = "AN_ID",
  ppm_mass_tol = 5,
  abs_mass_tol = 0,
  param_adduct = "[M+H]+",
  output_file = NULL,
  ...
)

CFM_fraggen(
  smiles_or_inchi = "[H]C1(O)O[C@]([H])(CO)[C@@]([H])(O)[C@]([H])(O)[C@@]1([H])O",
  max_depth = 2,
  param_adduct = "[M+H]+",
  output_file_or_dir = NULL
)

read_CFM_predict_result(
  result_path = "c:/Users/91879/OneDrive/Code/Docker/cfm/data/cfm_predict_result.txt"
)
}
\arguments{
\item{smiles_or_inchi_or_file}{smiles}

\item{prob_thresh}{prob}

\item{param_adduct}{adduct}

\item{annotate_fragments}{logic}

\item{output_file_or_dir}{path}

\item{apply_postproc}{logic}

\item{suppress_exceptions}{logic}

\item{smiles_or_inchi}{smiles}

\item{spectrum_file}{NULL}

\item{id}{ID}

\item{ppm_mass_tol}{20}

\item{abs_mass_tol}{0}

\item{output_file}{NULL}

\item{max_depth}{1}

\item{result_path}{path}
}
\value{
list of cfm data

cfm data

CFM_data

list of cfm data
}
\description{
see CFM

this function generate CFM_data by CFM_fraggen,
not rely on Spectra input

\link{CFM_predict} and then \link{CFM_annotate}
}
\section{Functions}{
\itemize{
\item \code{CFM_predict()}: predict

\item \code{CFM_annotate()}: anntate

\item \code{CFM_annotate_by_fraggen()}: fraggen and annotate

\item \code{CFM_annotate_by_predict()}: predict and annotate

\item \code{CFM_fraggen()}: fraggen

\item \code{read_CFM_predict_result()}: read_CFM_predict_result

}}
\note{
filter fragment of mz > 30 to avoid error when smiles transform to sdf
}
