% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_CFM.R
\name{CFM_annotate}
\alias{CFM_annotate}
\title{Title}
\usage{
CFM_annotate(
  smiles_or_inchi = "[H]C1(O)O[C@]([H])(CO)[C@@]([H])(O)[C@]([H])(O)[C@@]1([H])O",
  spectrum_file = NULL,
  id = "AN_ID",
  ppm_mass_tol = 5,
  abs_mass_tol = 0.01,
  param_adduct = "[M+H]+",
  output_file = NULL
)
}
\arguments{
\item{output_file}{}
}
\description{
Title
}
