% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_CFM.R
\name{CFM_annotate_by_fraggen}
\alias{CFM_annotate_by_fraggen}
\title{CFM_annotate_by_fraggen}
\usage{
CFM_annotate_by_fraggen(
  smiles_or_inchi = "[H]C1(O)O[C@]([H])(CO)[C@@]([H])(O)[C@]([H])(O)[C@@]1([H])O",
  spectrum_file = NULL,
  max_depth = 1,
  id = "AN_ID",
  ppm_mass_tol = 5,
  abs_mass_tol = 0,
  param_adduct = "[M+H]+",
  output_file = NULL,
  ...
)
}
\arguments{
\item{smiles_or_inchi}{smiles}

\item{spectrum_file}{NULL}

\item{max_depth}{1}

\item{id}{ID}

\item{ppm_mass_tol}{20}

\item{abs_mass_tol}{0}

\item{param_adduct}{adduct}

\item{output_file}{NULL}
}
\value{
CFM_data
}
\description{
this function generate CFM_data by CFM_fraggen,
not rely on Spectra input
}
\note{
filter fragment of mz > 30 to avoid error when smiles transform to sdf
}
