% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_CFM.R
\name{CFM_annotate_by_predict}
\alias{CFM_annotate_by_predict}
\title{CFM_annotate_by_predict}
\usage{
CFM_annotate_by_predict(
  smiles_or_inchi = "[H]C1(O)O[C@]([H])(CO)[C@@]([H])(O)[C@]([H])(O)[C@@]1([H])O",
  id = "AN_ID",
  ppm_mass_tol = 5,
  abs_mass_tol = 0,
  param_adduct = "[M+H]+",
  output_file = NULL,
  ...
)
}
\arguments{
\item{smiles_or_inchi}{smiles}

\item{id}{id}

\item{ppm_mass_tol}{ppm}

\item{abs_mass_tol}{mz tol}

\item{param_adduct}{adduct}

\item{output_file}{file path}
}
\description{
\link{CFM_predict} and then \link{CFM_annotate}
}
