% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_CFM.R
\name{CFM_annotate_isotopologues}
\alias{CFM_annotate_isotopologues}
\title{CFM_annotate_isotopologues}
\usage{
CFM_annotate_isotopologues(
  sp,
  cfmd,
  isotope = "[13]C",
  iso.count = 0,
  ppm = 20
)
}
\arguments{
\item{sp}{Spectra}

\item{cfmd}{cfmd_data}

\item{isotope}{"\link{13}C"}

\item{iso.count}{num}

\item{ppm}{20}
}
\value{
null
}
\description{
CFM_annotate_isotopologues
}
