% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_CFM.R
\name{cfm_data_get_fragment_group}
\alias{cfm_data_get_fragment_group}
\title{Title}
\usage{
cfm_data_get_fragment_group(cfm_data, ppm = 10)
}
\arguments{
\item{cfm_data}{cfmd}

\item{ppm}{10}
}
\value{
cfmd
}
\description{
Title
}
