% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_CFM.R
\name{CFM_data_get_igraph}
\alias{CFM_data_get_igraph}
\title{CFM_data_get_igraph
get cfm igraph and atom map}
\usage{
CFM_data_get_igraph(object)
}
\arguments{
\item{object}{cfmd}
}
\value{
cfmd
}
\description{
CFM_data_get_igraph
get cfm igraph and atom map
}
