% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_CFM.R
\name{CFM_predict}
\alias{CFM_predict}
\title{CFM predict}
\usage{
CFM_predict(
  smiles_or_inchi_or_file = "CCCNNNC(O)O",
  prob_thresh = 0.001,
  param_adduct = "[M+H]+",
  annotate_fragments = 1,
  output_file_or_dir = tempfile(),
  apply_postproc = 1,
  suppress_exceptions = 1
)
}
\arguments{
\item{suppress_exceptions}{}
}
\description{
CFM predict
}
