% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSdev-function.R
\name{checkSampleInfo}
\alias{checkSampleInfo}
\title{checkSampleInfo}
\usage{
checkSampleInfo(object)
}
\arguments{
\item{object}{a \code{MSdev} object}
}
\value{
a \code{MSdev} object
}
\description{
manually check sampleInfo using excel
}
