% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_Spectra.R
\name{combineSpectra_ce_max_precursor}
\alias{combineSpectra_ce_max_precursor}
\title{combineSpectra_ce_max_precursor}
\usage{
combineSpectra_ce_max_precursor(sp)
}
\arguments{
\item{sp}{sp}
}
\value{
sp
}
\description{
combineSpectra_ce_max_precursor
}
