% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_Spectra.R
\name{combineSpectra_groupby_ce}
\alias{combineSpectra_groupby_ce}
\title{combineSpectra_groupby_ce}
\usage{
combineSpectra_groupby_ce(sp, minProp = 0.5, ppm = 5, plot = T, ...)
}
\arguments{
\item{sp}{Spectra}

\item{minProp}{NUM}
}
\value{
Spectra
}
\description{
combineSpectra_groupby_ce
}
