% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_ms_condition.R
\name{create_ms_condition}
\alias{create_ms_condition}
\title{create_ms_condition}
\usage{
create_ms_condition()
}
\description{
create a set of ms param,
open an excel windows to edit param,
do not support modifying param terms right now,
then record in inner database (/data/ms_condition.Rda)
if MSC_id existed in database, auto replace.
}
\seealso{
\code{\link[=edit_ms_condition]{edit_ms_condition()}} to edit, \code{\link[=show_ms_condition]{show_ms_condition()}} to show
}
