% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_DEP.R
\name{DEP_add_rejections}
\alias{DEP_add_rejections}
\title{add_rejections_no_p.adj}
\usage{
DEP_add_rejections(data.se, p.adjust = T, lfc = 0.5)
}
\arguments{
\item{data.se}{SE}

\item{lfc}{fc threshold}

\item{alpha}{p threshold}
}
\value{
se
}
\description{
reference to DEP::add_rejections(),which not support significant with out p adjust,
this function as supplymentary
}
