% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_DEP.R
\name{DEP_test_diff}
\alias{DEP_test_diff}
\title{DEP.test.diff}
\usage{
DEP_test_diff(data.se, type = "all", ...)
}
\arguments{
\item{data.se}{SE}
}
\value{
SE
}
\description{
warpper of DEP::test_diff
}
