% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edit_ms_condition.R
\name{edit_ms_condition}
\alias{edit_ms_condition}
\title{edit_ms_condition}
\usage{
edit_ms_condition(MSC_id = "MSC0001")
}
\arguments{
\item{MSC_id}{\emph{character} or \emph{integer}, indicate MSC_id or its index}
}
\description{
open an excel window to edit ms_conditon,
MSC_id to indicate which record to edit,
remove record if delete MSC_id in excel
}
