% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand_adduct_from_compounds.R
\name{expand_adduct_from_compounds}
\alias{expand_adduct_from_compounds}
\title{Expand compound to adduct}
\usage{
expand_adduct_from_compounds(compound.record, polarity)
}
\arguments{
\item{polarity}{}
}
\description{
Using formula, exact.mass and polarity, expand adduct by enviPat::adduct
}
