% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_ggplot.R
\name{export_graph2pdf}
\alias{export_graph2pdf}
\title{export plot to pdf file
pdf() and export::graph2pdf() not support \code{append} arg
using qpdf::pdf_combine() to realize that function}
\usage{
export_graph2pdf(p, file_path, append = F, ...)
}
\arguments{
\item{p}{ggplot}

\item{file_path}{file path}

\item{append}{logic}

\item{...}{}
}
\value{
null
}
\description{
export plot to pdf file
pdf() and export::graph2pdf() not support \code{append} arg
using qpdf::pdf_combine() to realize that function
}
