% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_sample_information_from_wiff.R
\name{export_sample_information_from_wiff}
\alias{export_sample_information_from_wiff}
\title{Detect raw data and export sample information}
\usage{
export_sample_information_from_wiff(raw.data.dir, project.dir)
}
\arguments{
\item{raw.data.dir}{directory of raw data}

\item{project.dir}{directory to export}
}
\value{
ms.ana objective
}
\description{
Detect raw data (*.wiff and *.wiff.scan) from Sciex 6600
}
