% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_xcms.R
\name{extract_chrom}
\alias{extract_chrom}
\title{extract_chrom}
\usage{
extract_chrom(xcms.xcms, rtr, mzr, sample)
}
\arguments{
\item{xcms.xcms}{XCMSnExp object}

\item{rtr}{rt range}

\item{mzr}{MZ range}

\item{sample}{}
}
\value{
xcms
}
\description{
extract_chrom
}
