% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_spectra.R
\name{extract_spectra}
\alias{extract_spectra}
\title{Extract Spectra}
\usage{
extract_spectra(ms.ana)
}
\arguments{
\item{ms.ana}{}
}
\value{
ms.ana
}
\description{
Extract feature MSpectra from XCMSnExp and convert to Spectra
}
