% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSdev-function.R
\name{extractSpectra_fullscan_DDA}
\alias{extractSpectra_fullscan_DDA}
\title{extractSpectra_fullscan_DDA}
\usage{
extractSpectra_fullscan_DDA(object)
}
\arguments{
\item{object}{a \code{MSdev} object}
}
\value{
a \code{MSdev} object
}
\description{
extract all MS2 Spectra from \code{object@sampleInfo$msDataFile} which \code{sampleInfo$xcmsProcessing} \%in\% \% c("both","MS2"),
return store in \code{object@spectra$positiveMS2}
}
