% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSdev-function.R
\name{featureCandidate}
\alias{featureCandidate}
\title{featureCandidate}
\usage{
featureCandidate(
  object,
  mz.ppm = 20,
  spectraDatabase =
    "C:\\\\Users\\\\91879\\\\OneDrive\\\\Documents\\\\Code\\\\R\\\\Projecct\\\\2022.1.17_Compounds.database\\\\Spectra.integrated.database.integration.2022_02_12.Rdata"
)
}
\arguments{
\item{object}{a \code{MSdev} object}

\item{mz.ppm}{mz error}

\item{spectraDatabase}{databse path}
}
\value{
a \code{MSdev} object
}
\description{
match feature with database by mz, return all spectra matched in a list  splited by feature
}
