% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_xcms.R
\name{featureDefinitions_PeakSta}
\alias{featureDefinitions_PeakSta}
\title{featureDefinitions_PeakSta}
\usage{
featureDefinitions_PeakSta(xcms.xcms)
}
\arguments{
\item{xcms.xcms}{}
}
\description{
extract features' median rt, sn and maxo,
\code{xcms::featureDefinitions()} return a \code{DataFrame}, in which rtmin, rtmax, rtmed was median of \code{xcms::chromPeaks()$rt},
but not the median range of peaks
}
