% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSdev-function.R
\name{featureSpectra_fullscan_DDA}
\alias{featureSpectra_fullscan_DDA}
\title{featureSpectra_fullscan_DDA}
\usage{
featureSpectra_fullscan_DDA(object)
}
\arguments{
\item{object}{a \code{MSdev} object}
}
\value{
a \code{MSdev} object
}
\description{
extrat spectra from \code{MSdev@spectra} according to mz and rt of feature,
extracted spectra store in \code{object@spectra$positiveFeatureMS2} and \code{object@spectra$negativeFeatureMS2},
a list contain \code{Spectra} object of each feature, empty \code{Spectra} with precursorMz and rtime
}
