% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSdev-function.R
\name{findISMSdev}
\alias{findISMSdev}
\title{findISMSdev}
\usage{
findISMSdev(object, to.adjust = "featureRaw", corr.thred = 0.6)
}
\arguments{
\item{corr.thred}{}
}
\description{
find features of internals standard listed in \code{object@experimentInfo@Internal_Standard}
by \code{Exact_mass} and \code{Retention_time} (if provide),
only \link{M+H} and \link{M-H} are considered. Correlation and intensity will be plot based on \code{object@statData[["featureRaw"]]}, please check.
A column "internal_standard" will be added in \code{object@statData[["featureRaw"]]}
}
