% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_element_table.R
\name{fix_element_table}
\alias{fix_element_table}
\title{fix_element_table}
\usage{
fix_element_table()
}
\description{
this function add natural elemetn form, such as \link{12}C, \link{1}H to dataset from lc8 and enviPat
}
\note{
this will lead to warning in enviPat function and maybe unknown error, not apply
}
