% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_enviPat_lc8.R
\name{formula_mz_lc8}
\alias{formula_mz_lc8}
\title{formula_mz_lc8}
\usage{
formula_mz_lc8(Formula = "C2H4O1S2P1")
}
\arguments{
\item{Formula, }{such as \code{"C2H4"}}
}
\description{
calculate mz of a given formula
}
