% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_Chemmine.R
\name{get_atom_map}
\alias{get_atom_map}
\title{get_atom_map}
\usage{
get_atom_map(
  sdf.parent,
  sdf.product,
  ig.parent = get_sdf_igraph(sdf.parent),
  ig.product = get_sdf_igraph(sdf.product),
  iso_ele = "[13]C",
  return.type = c("most_prob", "prob_matrix")
)
}
\arguments{
\item{sdf.parent}{sdf}

\item{sdf.product}{sdf}

\item{ig.parent}{ig}

\item{ig.product}{ig}

\item{return.type}{string}
}
\value{
if prob_matrix, return a matrix, col: atom of product, row: atom of parent
if most_prob, return the vector of most likely map
}
\description{
map atom of two molecular
}
