% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_Spectra.R
\name{get_CFM_data_Spectra}
\alias{get_CFM_data_Spectra}
\title{get_CFM_data_Spectra}
\usage{
get_CFM_data_Spectra(cfmd)
}
\arguments{
\item{cfm.data}{from read_CFM_xxx}
}
\value{
Spcetra
}
\description{
get_CFM_data_Spectra
}
