% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_xcms.R
\name{get_chroms_data}
\alias{get_chroms_data}
\title{get_chroms_data}
\usage{
get_chroms_data(xchrom)
}
\arguments{
\item{xchrom}{XChromatograms}
}
\value{
xcms
}
\description{
extract chomatogram data to a data.frame
}
