% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DDA-function.R
\name{get_DDA_scan_stat}
\alias{get_DDA_scan_stat}
\title{get_DDA_scan_stat
stimulate DDA cycle and assign ms2 to feature, now just support single file}
\usage{
get_DDA_scan_stat(xcms.scan, feature_def, dynamic_time = 60)
}
\arguments{
\item{dynamic_time}{}
}
\description{
get_DDA_scan_stat
stimulate DDA cycle and assign ms2 to feature, now just support single file
}
