% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_xcms.R
\name{get_features_from_xcms}
\alias{get_features_from_xcms}
\title{get_features_from_xcms}
\usage{
get_features_from_xcms(xcms.xcms, missing = NA)
}
\arguments{
\item{xcms.xcms}{XCMSnExp object}
}
\value{
a SummarizedExperiment subject
}
\description{
extract feature data from xcms::XCMSnExp,
calculate RSD of QC and Sample
( note this rely on character "QC" and "Sample" in \code{sampleNames(xcms.xcms)} )
}
