% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_xcms.R
\name{get_intensity_rtime_df_from_XChromatogram}
\alias{get_intensity_rtime_df_from_XChromatogram}
\title{get_intensity_rtime_df_from_XChromatogram}
\usage{
get_intensity_rtime_df_from_XChromatogram(xchrom)
}
\arguments{
\item{xchrom}{}
}
\description{
extract chomatogram data to a data.frame
}
