% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSIP-function.R
\name{get_isotopologues_CFM_annotation}
\alias{get_isotopologues_CFM_annotation}
\title{get_isotopologues_CFM_annotation}
\usage{
get_isotopologues_CFM_annotation(iso.list)
}
\arguments{
\item{iso.list}{list of cfm data}
}
\value{
list of cfm data
}
\description{
get_isotopologues_CFM_annotation
}
