% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSIP-function.R
\name{get_isotopologues_Spectra_process}
\alias{get_isotopologues_Spectra_process}
\title{get_isotopologues_Spectra_process}
\usage{
get_isotopologues_Spectra_process(iso.list)
}
\arguments{
\item{iso.list}{iso.list}
}
\value{
iso.list
}
\description{
get_isotopologues_Spectra_process
}
