% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSdev-function.R
\name{get_MS_sampleinfo}
\alias{get_MS_sampleinfo}
\title{get_MS_sampleinfo}
\usage{
get_MS_sampleinfo(raw.data.dir, rawDataFormat = ".raw", verbose = T)
}
\arguments{
\item{object}{a \code{MSdev} object

grep "pos" and "neg" for ion mode

grep "blank", "blk" and "QC" for sample type, other samples will regard as "Sample"}
}
\value{
a \code{MSdev} object
}
\description{
read in ms raw data from \code{object@projectInfo$msDataDir}
and generate a table \code{sampleInfo}
note this function read in file according to their file names, those contaion both "pos" and "neg"
will be regarded as two files
}
