% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSIP-function.R
\name{get_MSdev_isotopologues}
\alias{get_MSdev_isotopologues}
\title{get_MSdev_isotopologues}
\usage{
get_MSdev_isotopologues(object)
}
\arguments{
\item{object}{MSdev}
}
\value{
a list of isotopologues
}
\description{
get_MSdev_isotopologues
}
