% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSdev-function.R
\name{get_MSdev_ms2_Spectra}
\alias{get_MSdev_ms2_Spectra}
\title{get_MSdev_ms2_Spectra}
\usage{
get_MSdev_ms2_Spectra(object)
}
\arguments{
\item{object}{MSdev}
}
\value{
Spectra
}
\description{
get_MSdev_ms2_Spectra
}
