% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_mzR.R
\name{get_MSinfo_mzR}
\alias{get_MSinfo_mzR}
\title{get_MSinfo_mzR
get MS experiment time by mzR}
\usage{
get_MSinfo_mzR(msDataFiles)
}
\arguments{
\item{msDataFiles}{file path}
}
\value{
mzR info
}
\description{
get_MSinfo_mzR
get MS experiment time by mzR
}
