% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_MetaboSignal.R
\name{get_node_info}
\alias{get_node_info}
\title{get_node_info}
\usage{
get_node_info(all_nodes)
}
\arguments{
\item{all_nodes}{kegg id, currently support cpd(compound) and hsa(gene)}
}
\description{
get_node_info
}
