% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_Chemmine.R
\name{get_smiles_sdf}
\alias{get_smiles_sdf}
\title{get_smile_sdf}
\usage{
get_smiles_sdf(smiles, smiles.id = NULL)
}
\arguments{
\item{smiles}{smiles}

\item{smiles.id}{NULL}
}
\value{
sdf
}
\description{
get_smile_sdf
}
