% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_Spectra.R
\name{get_Spectra_data}
\alias{get_Spectra_data}
\title{get_Spectra_data}
\usage{
get_Spectra_data(sp, var = c("precursorMz", "collisionEnergy"))
}
\arguments{
\item{sp}{\code{Spectra} object}

\item{var}{any variables of \code{Spectra::spectraVariables()}}
}
\value{
a dataframe with mz and intensity and vars
}
\description{
get_Spectra_data
}
