% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_Spectra.R
\name{get_Spectra_from_CFM_data}
\alias{get_Spectra_from_CFM_data}
\title{get_Spectra_from_CFM}
\usage{
get_Spectra_from_CFM_data(cfmd)
}
\arguments{
\item{cfm.data}{from read_CFM_xxx}
}
\value{
Spcetra
}
\description{
get_Spectra_from_CFM
}
