% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_Tuner.R
\name{get_xcms_Autotuner}
\alias{get_xcms_Autotuner}
\title{get xcms param by Autotuner}
\usage{
get_xcms_Autotuner(xcms.xcms)
}
\arguments{
\item{xcms.xcms}{xcms}
}
\value{
Autotuner
}
\description{
get xcms param by Autotuner
}
