% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_xcms.R
\name{get_xcms_chromatogram}
\alias{get_xcms_chromatogram}
\title{extract_chrom}
\usage{
get_xcms_chromatogram(object, ...)
}
\arguments{
\item{object}{XCMSnExp object}

\item{rt}{rt range}

\item{mz}{MZ range}
}
\value{
xcms
}
\description{
extract_chrom
}
