% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_xcms.R
\name{get_xcms_feature_ms1_candidate}
\alias{get_xcms_feature_ms1_candidate}
\title{match feature to database based on mz and rt
MSDB_id in db with mz error < mz.ppm will b}
\usage{
get_xcms_feature_ms1_candidate(
  xcms.xcms,
  cpdb,
  mz.ppm = 10,
  rt.tol = Inf,
  expand_adduct = F,
  ...
)
}
\arguments{
\item{xcms.xcms}{XCMSnExp object}

\item{mz.ppm}{num}

\item{rt.tol}{}
}
\value{
xcms
}
\description{
match feature to database based on mz and rt
MSDB_id in db with mz error < mz.ppm will b
}
