% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_xcms.R
\name{get_xcms_scan_feature_id}
\alias{get_xcms_scan_feature_id}
\title{get_xcms_scan_feature_id}
\usage{
get_xcms_scan_feature_id(xcms.scan, featuredef)
}
\arguments{
\item{featuredef}{}
}
\description{
match xcms scan of msLevel 2 to feature Definitions
based on precursorMZ, retentionTime in \code{xcms.scan};
peakRtMin, peakRtMax, feature_id in \code{featuredef}
}
