% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_mzR.R
\name{getmsExpTime}
\alias{getmsExpTime}
\title{getmsExpTime}
\usage{
getmsExpTime(msDataFiles)
}
\arguments{
\item{msdataFiles}{}
}
\value{
a data.frame with two column: msDataFiles, ExpTime
}
\description{
get MS experiment time by mzR::runInfo()$startTimeStamp
}
