% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSdev-function.R
\name{getSEMSdev}
\alias{getSEMSdev}
\title{getSEMSdev}
\usage{
getSEMSdev(MSdev.obj)
}
\arguments{
\item{MSdev.obj}{}
}
\description{
extrat SummarizedExperiment::SummarizedExperiment, which combine coldata(sample info) and rowdata (metabolites/feature)
and perform data filter, normalization and imputation, refer to package DEP
}
