% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSdev-function.R
\name{getStaDataMSdev}
\alias{getStaDataMSdev}
\title{getStaDataMSdev}
\usage{
getStaDataMSdev(
  object,
  missing = NA,
  MSDB.keys = c("name", "adduct", "formula", "inchikey", "database_origin")
)
}
\arguments{
\item{MSDB.keys}{}
}
\description{
getStaDataMSdev
}
