% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_ggplot.R
\name{ggplot.sum.patchwork}
\alias{ggplot.sum.patchwork}
\title{ggplot.sum.patchwork}
\usage{
\method{ggplot}{sum.patchwork}(ggplot.list)
}
\arguments{
\item{ggplot.list}{a list with all item as ggplot objective}
}
\description{
add all ggplot by patchwork
}
