% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_caret.R
\name{ggplot_roc}
\alias{ggplot_roc}
\title{ggplot_roc}
\usage{
ggplot_roc(roc.train, roc.test)
}
\arguments{
\item{roc.train}{roc}

\item{roc.test}{roc}
}
\value{
ggplot
}
\description{
plot ROC of train and test
}
