% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_ggplot.R
\name{ggplot_sum_patchwork}
\alias{ggplot_sum_patchwork}
\title{ggplot_sum_patchwork}
\usage{
ggplot_sum_patchwork(ggplot.list)
}
\arguments{
\item{ggplot.list}{a list with all item as ggplot objective}
}
\description{
add all ggplot by patchwork
}
