% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_string.R
\name{groupStringFactor}
\alias{groupStringFactor}
\title{groupStringFactor}
\usage{
groupStringFactor(x)
}
\arguments{
\item{x}{str}
}
\value{
str with factor
}
\description{
convert a vector to a factor,
levels will be ordered by 1: string, "con" or "wt" will be placed in first;
2: others will be ordered according to0 number
}
