% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_tidyverse.R
\name{list2df}
\alias{list2df}
\title{list2df}
\usage{
list2df(x)
}
\arguments{
\item{x}{}
}
\value{
data.frame
}
\description{
convert list to data.frame. In each list,
sub-list will be replaced with "large list" and missing value will be fill by NA.
See \code{data.table::rbindlist}
}
