% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_base.R
\name{load_as_var}
\alias{load_as_var}
\title{load_as_var}
\usage{
load_as_var(file_to_load)
}
\arguments{
\item{file_to_load}{}
}
\description{
load Rdata file and return, only one variable in \code{file_to_load}
}
