% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_adduct_with_eicSimilarity.R
\name{match_adduct_with_eicSimilarity}
\alias{match_adduct_with_eicSimilarity}
\title{match_adduct_with_eicSimilarity}
\usage{
match_adduct_with_eicSimilarity(MS.network, xcms.xcms)
}
\arguments{
\item{xcms.xcms}{}
}
\description{
extract feature's chromatogram and compare their similarity to main peak
}
