% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Isotopologues-function.R
\name{match_isotopes_to_xcms_feature}
\alias{match_isotopes_to_xcms_feature}
\title{match_isotopes_to_xcms_feature}
\usage{
match_isotopes_to_xcms_feature(
  isotopes.network,
  xcms.xcms,
  ppm.thresh = 20,
  rt.tol = 10,
  value = "ratio_sub_nature"
)
}
\arguments{
\item{ppm.thresh}{ppm.thresh}

\item{MS.network}{MS.network}

\item{xcms.features}{xcms.features}
}
\value{
xcms
}
\description{
match_isotopes_to_xcms_feature
}
