% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_math.R
\name{match_mz_rt}
\alias{match_mz_rt}
\title{match 2 list of ion based on mz and rt
return all matched, include multiple match}
\usage{
match_mz_rt(
  mz1,
  rt1 = rep(NA, length(mz1)),
  mz2,
  rt2 = rep(NA, length(mz2)),
  mz.ppm = 10,
  rt.tol = Inf
)
}
\arguments{
\item{mz1}{num}

\item{rt1}{num}

\item{mz2}{num}

\item{rt2}{num}

\item{mz.ppm}{num}

\item{rt.tol}{num}
}
\description{
match 2 list of ion based on mz and rt
return all matched, include multiple match
}
