% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_base.R
\name{matrixSub}
\alias{matrixSub}
\title{matrixSub}
\usage{
matrixSub(v1, v2)
}
\arguments{
\item{v1}{vector}

\item{v2}{vector}
}
\value{
matrix
}
\description{
expand 2 vector to matrix and "\code{-}"
}
\examples{
a <- 3:8
b <- 1:2
matrixSub(a,b)
}
