% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_math.R
\name{median_part}
\alias{median_part}
\title{median_part}
\usage{
median_part(x, n = 10)
}
\arguments{
\item{x}{vector}

\item{n}{number of element}
}
\value{
vector
}
\description{
similar to head() and tail(), return the median of a vector
}
