% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_MetaboSignal.R
\name{MetaboSignalNetworkTable2df}
\alias{MetaboSignalNetworkTable2df}
\title{MetaboSignalNetworkTable2df}
\usage{
MetaboSignalNetworkTable2df(network_table)
}
\arguments{
\item{network_table}{matrix with 3 column: source, target, interaction_type}
}
\description{
MetaboSignalNetworkTable2df
}
