% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_MetaboSignal.R
\name{MetaboSignalNetworkTableStandardNodeName}
\alias{MetaboSignalNetworkTableStandardNodeName}
\title{MetaboSignalNetworkTableStandardNodeName}
\usage{
MetaboSignalNetworkTableStandardNodeName(network_table, add_hsa = T)
}
\arguments{
\item{add_hsa}{}
}
\description{
add or remove "hsa:" from network_table
}
