% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Molecule_igraph-functions.R
\name{Molecule_igraph_add_isotopomer}
\alias{Molecule_igraph_add_isotopomer}
\title{Add isotopomer}
\usage{
Molecule_igraph_add_isotopomer(
  Molecule_igraph,
  isotopomer = NULL,
  iso_vec = NULL,
  abundance = NA,
  path = NA
)
}
\arguments{
\item{Molecule_igraph}{\code{Molecule_igraph}}

\item{isotopomer}{isotopomer name}

\item{iso_vec}{isotope of atom, such as c("C_1" = "\link{13}C")}

\item{abundance}{number}
}
\value{
\code{Molecule_igraph}
}
\description{
Add isotopomer
}
\section{Functions}{
\itemize{
\item \code{Molecule_igraph_add_isotopomer()}: Add isotopomer

}}
