% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MS_exp-class.R
\docType{class}
\name{MS_Exp-class}
\alias{MS_Exp-class}
\title{MS_Exp-class}
\description{
MS_Exp-class
}
\section{Slots}{

\describe{
\item{\code{General}}{tbl.}

\item{\code{Pre_process}}{list.}

\item{\code{Moblie_phase_A}}{list.}

\item{\code{Moblie_phase_B}}{list.}

\item{\code{Chroma_column}}{list.}

\item{\code{Chroma_gradient}}{list.}

\item{\code{Mass_Spectrum}}{list.}
}}

