% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_proteoWizard.R
\name{msconvert_raw2mzML}
\alias{msconvert_raw2mzML}
\title{msconvert_raw2mzML}
\usage{
msconvert_raw2mzML(
  raw.files,
  mzML.files,
  BPPARAM = SnowParam(workers = parallel::detectCores() - 1)
)
}
\arguments{
\item{raw.files}{file path}

\item{mzML.files}{file path}

\item{BPPARAM}{biocparall}
}
\value{
null
}
\description{
msconvert_raw2mzML
}
