% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSdev-function.R
\name{MSdev_add_sample}
\alias{MSdev_add_sample}
\title{MSdev_add_sample}
\usage{
MSdev_add_sample(object, raw.data.dir = object@projectInfo$rawDataDir)
}
\arguments{
\item{object}{MSdev}

\item{raw.data.dir}{file path}
}
\value{
MSdev
}
\description{
MSdev_add_sample
}
