% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSdev1.3-Funcion.R
\name{MSdev_annotation}
\alias{MSdev_annotation}
\title{MSdev annotation}
\usage{
MSdev_annotation(object, db.path = "d:/MSdb/msdb.HMDB.Rdata", ...)
}
\arguments{
\item{object}{MSdev}

\item{db.path}{CompoundDB}

\item{...}{}
}
\description{
MSdev annotation
}
