% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSdev-function.R
\name{MSdev_checkSampleInfo}
\alias{MSdev_checkSampleInfo}
\title{check SampleInfo in excel}
\usage{
MSdev_checkSampleInfo(object)
}
\arguments{
\item{object}{a \code{MSdev} object}
}
\value{
MSdev a \code{MSdev} object
}
\description{
manually check sampleInfo using excel
}
