% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSdev-function.R
\name{MSdev_extract_Spectra}
\alias{MSdev_extract_Spectra}
\title{MSdev_extract_Spectra}
\usage{
MSdev_extract_Spectra(object, eval.noise = F, eval.ms1 = F)
}
\arguments{
\item{object}{MSdev}
}
\value{
MSdev
}
\description{
Extract all spectra, split to MS1 and MS2, store as \code{onDiskData}
}
