% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSdev-function.R
\name{MSdev_get_feature_chrom}
\alias{MSdev_get_feature_chrom}
\title{MSdev_get_feature_chrom}
\usage{
MSdev_get_feature_chrom(object, feature.list = NULL)
}
\arguments{
\item{object}{MSdev}

\item{feature.list}{list of feature_id, named "Positve" and "Negative"}
}
\value{
MSdev
}
\description{
MSdev_get_feature_chrom
}
