% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSdev-function.R
\name{MSdev_get_SE}
\alias{MSdev_get_SE}
\title{getSEMSdev}
\usage{
MSdev_get_SE(MSdev.obj)
}
\arguments{
\item{MSdev.obj}{MSdev}
}
\value{
MSdev
}
\description{
extrat SummarizedExperiment::SummarizedExperiment, which combine coldata(sample info) and rowdata (metabolites/feature)
and perform data filter, normalization and imputation, refer to package DEP
}
