% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSdev-function.R
\name{MSdev_get_Stat}
\alias{MSdev_get_Stat}
\title{MSdev_get_Stat}
\usage{
MSdev_get_Stat(
  object,
  QC_RSD = 0.3,
  keys = c("name", "formula", "kegg_id", "inchikey", "lipidclass")
)
}
\arguments{
\item{object}{MSdev}

\item{QC_RSD}{QC RSD thresh}
}
\value{
MSdev
}
\description{
extract data from xcms
retrieve compound info from CompDB
filter data
generate data.se
}
