% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSdev-function.R
\name{MSdev_match_Spectra_to_feature}
\alias{MSdev_match_Spectra_to_feature}
\title{MSdev_match_Spectra_to_feature}
\usage{
MSdev_match_Spectra_to_feature(object)
}
\arguments{
\item{object}{MSdev}
}
\value{
MSdev
}
\description{
MSdev_match_Spectra_to_feature
}
