% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSdev-function.R
\name{MSdev_xcms_group_features}
\alias{MSdev_xcms_group_features}
\title{MSdev_xcms_group_features}
\usage{
MSdev_xcms_group_features(object, diffRt = 5, intCor = 0.5, eicCor = 0.3, ...)
}
\arguments{
\item{object}{MSdev}

\item{diffRt}{30}

\item{intCor}{NULL}

\item{eicCor}{NULL}

\item{...}{MSdev}
}
\value{
MSdev
}
\description{
MSdev_xcms_group_features
}
