% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSdev-function.R
\name{MSdev_xcmsProcessing}
\alias{MSdev_xcmsProcessing}
\title{MSdev_xcmsProcessing}
\usage{
MSdev_xcmsProcessing(object, ...)
}
\arguments{
\item{object}{MSdev}

\item{...}{}
}
\value{
MSdev
}
\description{
MSdev_xcmsProcessing
}
