% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSIP-function.R, R/dev_CFM.R
\name{MSIP_merge_isotopologues}
\alias{MSIP_merge_isotopologues}
\alias{CFM_annotate_isotopologues}
\title{CFM_annotate_isotopologues}
\usage{
MSIP_merge_isotopologues(object, rt.tol = 5)

CFM_annotate_isotopologues(
  sp,
  cfmd,
  iso_ele = "[13]C",
  iso_count = 0,
  ppm = 20
)
}
\arguments{
\item{object}{MSIP/MSdev object}

\item{rt.tol}{permitted rt error to merge pos and neg data}

\item{sp}{Spectra}

\item{cfmd}{cfmd_data}

\item{iso_count}{num}

\item{ppm}{20}

\item{isotope}{"\link{13}C"}
}
\value{
MSIP/MSdev object

null
}
\description{
CFM_annotate_isotopologues
}
\section{Functions}{
\itemize{
\item \code{MSIP_merge_isotopologues()}: merge msip data of positive and negative

\item \code{CFM_annotate_isotopologues()}: CFM_annotate_isotopologues

}}
