% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSIP-function.R
\name{MSIP_get_isotopologues_CFM_annotation}
\alias{MSIP_get_isotopologues_CFM_annotation}
\title{get_isotopologues_CFM_annotation}
\usage{
MSIP_get_isotopologues_CFM_annotation(
  object,
  ppm = 20,
  BPPARAM = SnowParam(progressbar = T)
)
}
\arguments{
\item{object}{msdev}
}
\value{
list of cfm data
}
\description{
get_isotopologues_CFM_annotation
}
