% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSIP-function.R
\name{MSIP_get_isotopologues_data}
\alias{MSIP_get_isotopologues_data}
\title{get_MSdev_isotopologues}
\usage{
MSIP_get_isotopologues_data(object, iso_ele = "[13]C")
}
\arguments{
\item{object}{MSdev}
}
\value{
a list of isotopologues
}
\description{
extract iso-labeled compound info and spectra,
filter compound
}
