% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shiny_MSIP_SERVER.R
\name{MSIP_shiny_Acq_server}
\alias{MSIP_shiny_Acq_server}
\title{MSIP_shiny_Acq_server}
\usage{
MSIP_shiny_Acq_server(object)
}
\arguments{
\item{object}{MSdev}
}
\value{
MSdev
}
\description{
MSIP_shiny_Acq_server
}
