% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSIPCoreData-function.R
\name{heatmap_MSIPFragmentMap}
\alias{heatmap_MSIPFragmentMap}
\title{heatmap_MSIPFragmentMap}
\usage{
heatmap_MSIPFragmentMap(MSIPFragmentMap)
}
\arguments{
\item{MSIPFragmentMap}{MSIPFragmentMap}
}
\value{
heatmap
}
\description{
heatmap_MSIPFragmentMap
}
\section{Functions}{
\itemize{
\item \code{heatmap_MSIPFragmentMap()}: heatmap fragment map

}}
