% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_ggplot.R
\name{open.ggplot.win}
\alias{open.ggplot.win}
\title{open.ggplot.win}
\usage{
\method{open}{ggplot.win}(p, width = NA, height = NA)
}
\arguments{
\item{p}{ggplot objective}

\item{height}{}
}
\description{
create a temp.png file and open in Windows
}
