% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_base.R
\name{open_dir}
\alias{open_dir}
\title{open_dir
open dir by Windows}
\usage{
open_dir(x = getwd())
}
\arguments{
\item{dir}{a path of dir or file}
}
\value{
null
}
\description{
open_dir
open dir by Windows
}
