% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_ggplot.R
\name{open_ggplot_win}
\alias{open_ggplot_win}
\title{open_ggplot_win}
\usage{
open_ggplot_win(p, width = NA, height = NA)
}
\arguments{
\item{p}{ggplot objective}

\item{height}{}
}
\description{
create a temp.png file and open in Windows
}
