% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_ggplot.R
\name{open_plot_win}
\alias{open_plot_win}
\title{open_plot_win}
\usage{
open_plot_win(p, width = 5, height = 4)
}
\arguments{
\item{p}{ggplot/Complexheatmap}

\item{width}{num}

\item{height}{num}
}
\value{
null
}
\description{
create a temp.png file and open in Windows
}
