% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSdev-function.R
\name{plot_MSdev_feature_spectrum}
\alias{plot_MSdev_feature_spectrum}
\title{Title}
\usage{
plot_MSdev_feature_spectrum(MSdev.obj, feature.id)
}
\arguments{
\item{MSdev.obj}{MSdev}

\item{feature_id}{feature_id}
}
\value{
ggplot
}
\description{
Title
}
