% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatisticFunction.R
\name{plot_PCA}
\alias{plot_PCA}
\title{plotPCA}
\usage{
plot_PCA(pca.matrix, pca.group, showlabel = F)
}
\arguments{
\item{pca.matrix}{pca.matrix}

\item{pca.group}{pca.group}

\item{showlabel}{showlabel}
}
\value{
ggplot
}
\description{
plotPCA
}
