% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_Spectra.R
\name{plot_Spectra}
\alias{plot_Spectra}
\title{plot_Spectra}
\usage{
plot_Spectra(sp, label.top = 10)
}
\arguments{
\item{sp}{Spectra}

\item{label.top}{peaks to label}
}
\value{
ggplot
}
\description{
plot_Spectra
}
