% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_xcms.R
\name{plot_XChromatograms}
\alias{plot_XChromatograms}
\title{plot_XChromatograms}
\usage{
plot_XChromatograms(
  xchroms,
  norm = T,
  move = T,
  color_by = c("column", "row"),
  color_f = NULL,
  label_df = NULL
)
}
\arguments{
\item{move}{}
}
\value{
xcms
}
\description{
plot_XChromatograms
}
