% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_xcms.R
\name{plot_xcms_feature_chromatogram}
\alias{plot_xcms_feature_chromatogram}
\title{plot_xcms_feature_chromatogram}
\usage{
plot_xcms_feature_chromatogram(xcms.xcms, feature.id, sampleNames = NULL)
}
\arguments{
\item{xcms.xcms}{XCMSnExp object}

\item{feature.id}{feature id}

\item{sampleNames}{}
}
\value{
xcms
}
\description{
extract Chromatogram from xcms according to feature's mz range and plot
}
