% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_xcms.R
\name{plot_xcms_feature_intensity}
\alias{plot_xcms_feature_intensity}
\title{plot_xcms_feature_intensity}
\usage{
plot_xcms_feature_intensity(xcms.xcms, feature_id_to_show)
}
\arguments{
\item{feature_id_to_show}{}
}
\description{
plot feature's intensity, ordered by \code{pData(xcms.xcms)$analysis.time.positive} or
\code{pData(xcms.xcms)$analysis.time.negative}
}
