% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_xcms.R
\name{plot_xcms_features_distribution}
\alias{plot_xcms_features_distribution}
\title{plot_xcms_peaks_distribution}
\usage{
plot_xcms_features_distribution(
  xcms.xcms,
  plot.title = "Features distribution"
)
}
\arguments{
\item{plot.title}{}
}
\description{
plot_xcms_peaks_distribution
}
