% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_xcms.R
\name{plot_xcms_peaks_distribution}
\alias{plot_xcms_peaks_distribution}
\title{plot_xcms_peaks_distribution}
\usage{
plot_xcms_peaks_distribution(
  xcms.xcms,
  plot.title = "Peaks distribution",
  type = "o"
)
}
\arguments{
\item{type}{\code{"o"}, for geom_point, \code{"l"}, for geom_segment}
}
\description{
export peaks data by xcms::chromPeaks and plot by ggplot2
}
