% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_xcms.R
\name{plot_xcms_peaks_ms1_scans}
\alias{plot_xcms_peaks_ms1_scans}
\title{plot_xcms_peaks_ms1_scans}
\usage{
plot_xcms_peaks_ms1_scans(xcms.xcms, plot.title = "Peaks Sans of MS1")
}
\arguments{
\item{xcms.xcms}{should be a \code{XCMSnExp} object after \code{findChromPeaks}}

\item{plot.title}{}
}
\description{
plot scans number of MS1 levels in each peak, note that to many peaks will lead to stack,
apply \code{filterFile} to decrease peaks count
}
