% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_xcms.R
\name{plot_xcms_peaks_ms2_scans}
\alias{plot_xcms_peaks_ms2_scans}
\title{Title}
\usage{
plot_xcms_peaks_ms2_scans(xcms.xcms, plot.title = "Peaks Sans of MS2")
}
\arguments{
\item{xcms.xcms}{XCMSnExp object}

\item{plot.title}{}
}
\description{
Title
}
