% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_Spectra.R
\name{plotly_Spectra}
\alias{plotly_Spectra}
\title{plotly_Spectra}
\usage{
plotly_Spectra(sp, label.top = 10)
}
\arguments{
\item{sp}{Spectra}

\item{label.top}{highlight}
}
\value{
plotly
}
\description{
plotly_Spectra
}
