% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_Spectra.R
\name{plotly_Spectra_iso_mirror}
\alias{plotly_Spectra_iso_mirror}
\title{plotly_Spectra_iso_mirror}
\usage{
plotly_Spectra_iso_mirror(
  sp,
  sp.iso,
  ppm = 10,
  iso_mass_diff = 1.003355,
  iso_count = 1
)
}
\arguments{
\item{sp}{sp}

\item{sp.iso}{iso sp}

\item{ppm}{ppm}

\item{iso_mass_diff}{iso mass diff to nature}

\item{iso_count}{iso count}
}
\value{
plotly
}
\description{
plotly_Spectra_iso_mirror
}
