% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_Spectra.R
\name{plotly_Spectra_mirror}
\alias{plotly_Spectra_mirror}
\title{plotly_Spectra_mirror}
\usage{
plotly_Spectra_mirror(sp1, sp2)
}
\arguments{
\item{sp1}{sp1}

\item{sp2}{sp2}
}
\value{
plotly
}
\description{
plotly_Spectra_mirror
}
