% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatisticFunction.R
\name{plotPCA}
\alias{plotPCA}
\title{plotPCA}
\usage{
plotPCA(pca.matrix, pca.group, showlabel = F)
}
\arguments{
\item{pca.matrix}{pca.matrix}

\item{pca.group}{pca.group}

\item{showlabel}{showlabel}
}
\value{
ggplot
}
\description{
plotPCA
}
