% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSdev-function.R
\name{readInRawData}
\alias{readInRawData}
\title{readInRawData}
\usage{
readInRawData(object)
}
\arguments{
\item{object}{a \code{MSdev} object}
}
\value{
a \code{MSdev} object
}
\description{
read in ms raw data from \code{object@projectInfo$msDataDir}
and generate a table \code{sampleInfo}
note this function read in file according to their file names
}
\details{
discriminate sample type and ion mode according char in file names

grep "pos" and "neg" for ion mode

grep "blank", "blk" and "QC" for sample type, other samples will regard as "Sample"
}
