% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Retrieve_compounds_data_from_pubchem.R
\name{Retrieve_compounds_data_from_pubchem}
\alias{Retrieve_compounds_data_from_pubchem}
\title{Retrieve compound data}
\usage{
Retrieve_compounds_data_from_pubchem(compound.record.file)
}
\value{
compound.record
}
\description{
Retrieve compound data
}
