% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_RStudio.R
\name{setRStudioDir}
\alias{setRStudioDir}
\title{setRStudioDir}
\usage{
setRStudioDir(path = rstudioapi::getSourceEditorContext()$path)
}
\arguments{
\item{path}{path}
}
\value{
null
}
\description{
setRStudioDir
}
