% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_RStudio.R
\name{setwdActivedFilePath}
\alias{setwdActivedFilePath}
\title{setwdActivedFilePath}
\usage{
setwdActivedFilePath()
}
\value{
null
}
\description{
Set Working Directory as currently opened file in RStudio editor,
see \code{rstudioapi::getSourceEditorContext()}
}
