% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_ms_condition.R
\name{show_ms_condition}
\alias{show_ms_condition}
\title{show_ms_condition}
\usage{
show_ms_condition(MSC_id = "all")
}
\arguments{
\item{MSC_id}{\emph{character} or \emph{integer}, indicate MSC_id or its index}
}
\description{
show indicated ms_condition record,
or return all record
}
