% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_Spectra.R
\name{Spectra_filter_noise}
\alias{Spectra_filter_noise}
\title{Spectra_filter_noise}
\usage{
Spectra_filter_noise(sp)
}
\arguments{
\item{sp}{Spectra}
}
\value{
Spectra
}
\description{
Spectra_filter_noise
}
