% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_Spectra.R
\name{Spectra_get_noise}
\alias{Spectra_get_noise}
\title{Spectra_get_noise}
\usage{
Spectra_get_noise(sp)
}
\arguments{
\item{sp}{Spectra}
}
\value{
Spectra
}
\description{
estimate noise with max density
}
