% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_Spectra.R
\name{Spectra_get_purity}
\alias{Spectra_get_purity}
\title{Spectra_get_purity}
\usage{
Spectra_get_purity(sp, msLevel = 2, sp.ms1 = NULL)
}
\arguments{
\item{sp}{Spectra}
}
\value{
Spectra
}
\description{
estimate a ms2 spectra purity with MS1 scan
}
