% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_base.R
\name{split_df}
\alias{split_df}
\title{split_df
random split data.frame}
\usage{
split_df(df, n = 2)
}
\arguments{
\item{df}{data.frame}

\item{n}{number}
}
\value{
list
}
\description{
split_df
random split data.frame
}
