% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_string.R
\name{str_short}
\alias{str_short}
\title{str_short}
\usage{
str_short(x, n = 10)
}
\arguments{
\item{x}{str}

\item{n}{char number}
}
\value{
str
}
\description{
cut string with length >n and replace char > n with "..."
}
