% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_ggplot.R
\name{sum.patchwork}
\alias{sum.patchwork}
\title{sum.patchwork}
\usage{
\method{sum}{patchwork}(ggplot.list)
}
\arguments{
\item{ggplot.list}{a list with all item as ggplot objective}
}
\description{
add all ggplot by patchwork
}
