% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Molecule_vis.R
\name{vis_sdf_igraph}
\alias{vis_sdf_igraph}
\title{vis_sdf_igraph}
\usage{
vis_sdf_igraph(sdf.igraph, show_id = F, ...)
}
\arguments{
\item{sdf.igraph}{igraph}

\item{show_id}{logic}

\item{highlight}{vector}
}
\value{
vis html
}
\description{
vis_sdf_igraph
}
