% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_Chemmine.R
\name{vis_smiles}
\alias{vis_smiles}
\title{vis_smiles}
\usage{
vis_smiles(smiles, show.formula = T, show_id = T, highlight = NULL)
}
\arguments{
\item{smiles}{smiles}

\item{show.formula}{logic}

\item{highlight}{v}

\item{show.label}{logic}
}
\value{
vis
}
\description{
vis_smiles
}
