% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_openxlsx.R
\name{write.xlsx}
\alias{write.xlsx}
\title{write.xlsx}
\usage{
write.xlsx(data, file.dir)
}
\arguments{
\item{data}{data to write}

\item{file.dir}{file path}
}
\value{
null
}
\description{
write xlsx, dir will be created if not existed
}
