% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_xcms.R
\name{XChromatograms_fill_2point}
\alias{XChromatograms_fill_2point}
\title{XChromatograms_fill_2point}
\usage{
XChromatograms_fill_2point(xchroms)
}
\arguments{
\item{xchroms}{XChromatograms}
}
\value{
xcms
}
\description{
when xcms::findChromPeaks(), if any Chromatogram contain less than 2 point, this will lead to error
}
