% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_xcms.R
\name{XChromatograms_rt_unit}
\alias{XChromatograms_rt_unit}
\title{XChromatograms_rt_unit}
\usage{
XChromatograms_rt_unit(xchroms, unit_to = "s")
}
\arguments{
\item{xchroms}{\code{XChromatograms} or \code{MChromatograms} object}

\item{unit_to}{"s" or "m", "s": rtime*60; "m": rtime/60}
}
\value{
xcms
}
\description{
change rtime units, in some situation (such as SRM data from Thermo), rtime are recorded with unit "m",
this will lead to error when findChrompeaks
}
