% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DDA-function.R
\name{xcms_get_dda_ms2_assignment}
\alias{xcms_get_dda_ms2_assignment}
\title{xcms_get_dda_ms2_assignment}
\usage{
xcms_get_dda_ms2_assignment(xcms.xcms)
}
\arguments{
\item{featuredef}{}
}
\description{
match xcms scan of msLevel 2 to feature Definitions
based on precursorMZ, retentionTime in \code{xcms.scan};
peakRtMin, peakRtMax, feature_id in \code{featuredef}
}
