% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DDA-function.R
\name{xcms_get_dda_scan_stimulate}
\alias{xcms_get_dda_scan_stimulate}
\title{xcms_get_dda_scan_stimulate
stimulate DDA cycle and assign ms2 to feature, now just support single file}
\usage{
xcms_get_dda_scan_stimulate(xcms.xcms, dynamic_time = 60)
}
\arguments{
\item{dynamic_time}{}
}
\description{
xcms_get_dda_scan_stimulate
stimulate DDA cycle and assign ms2 to feature, now just support single file
}
