% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_xcms.R
\name{xcms_get_feature_def_stat}
\alias{xcms_get_feature_def_stat}
\title{stat featureDefinitions based on chrompeaks}
\usage{
xcms_get_feature_def_stat(xcms.xcms)
}
\arguments{
\item{xcms.xcms}{XCMSnExp object}
}
\description{
extract features' median rt, sn and maxo,
\code{xcms::featureDefinitions()} return a \code{DataFrame}, in which rtmin, rtmax, rtmed was median of \code{xcms::chromPeaks()$rt},
but not the median range of peaks. peakRtMin, peakRtMax, peakSN, peakMaxo are median of all peaks in a feature
}
