% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_xcms.R
\name{xcms_get_feature_group}
\alias{xcms_get_feature_group}
\title{xcms feature group}
\usage{
xcms_get_feature_group(xcms.xcms, diffRt = 5, intCor = 0.5, eicCor = 0.5)
}
\arguments{
\item{xcms.xcms}{XCMSnExp object}
}
\value{
xcms
}
\description{
xcms feature group
}
