% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_xcms.R
\name{xcms_get_feature_isotope_label}
\alias{xcms_get_feature_isotope_label}
\title{xcms_get_feature_isotope_label}
\usage{
xcms_get_feature_isotope_label(xcms.xcms, iso_ele = "[13]C", ...)
}
\arguments{
\item{xcms.xcms}{xcms}

\item{iso_ele}{"[13]C"}

\item{...}{}
}
\description{
A short description...
}
