% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_xcms.R
\name{xcms_get_feature_isotopologues}
\alias{xcms_get_feature_isotopologues}
\title{xcms_get_feature_isotopologues}
\usage{
xcms_get_feature_isotopologues(
  xcms.xcms,
  iso_ele = "[13]C",
  max_label = 10,
  ppm = 5,
  rt.tol = 5,
  net.degree.ratio = 0.5
)
}
\arguments{
\item{xcms.xcms}{xcms}

\item{max_label}{10}

\item{ppm}{20}

\item{net.degree.ratio}{0.5}

\item{isotope}{"[13]C"}
}
\value{
xcms
}
\description{
screen isotopologues peaks by mz and rt diff,
record result in featureDefinitions
}
