% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_xcms.R
\name{xcms_get_feature_ms1_candidate}
\alias{xcms_get_feature_ms1_candidate}
\title{match feature to database based on mz and rt
MSDB_id in db with mz error < mz.ppm will b}
\usage{
xcms_get_feature_ms1_candidate(
  xcms.xcms,
  cpdb,
  mz.ppm = 10,
  rt.tol = Inf,
  selected_adduct = MSCC::adduct.table$Adduct,
  ...
)
}
\arguments{
\item{xcms.xcms}{XCMSnExp object}

\item{cpdb}{compoundDb}

\item{mz.ppm}{num}

\item{rt.tol}{num}
}
\value{
xcms
}
\description{
match feature to database based on mz and rt
MSDB_id in db with mz error < mz.ppm will b
}
