% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xcms_processing.R
\name{xcms_processing}
\alias{xcms_processing}
\title{xcms_procecing}
\usage{
xcms_processing(sample.info, ion_mode)
}
\description{
read in files in sample.info$mzML.files. note the sample were reordered in xcms object by analysis time, may not identical to sample.name, should be correspondended by mzML.file
}
