% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_xcms.R
\name{xcmsProcessingMS1}
\alias{xcmsProcessingMS1}
\title{xcmsProcessingMS1}
\usage{
xcmsProcessingMS1(
  msDataFiles,
  ion_mode = NA,
  peaksGroup = NA,
  centWaveParam = xcms::CentWaveParam(ppm = 20, peakwidth = c(5, 30), snthresh = 10,
    prefilter = c(3, 100))
)
}
\arguments{
\item{msDataFiles}{\code{char} ms file (full) paths}

\item{ion_mode}{to filter ion_mode, 1: positive, 0: negative, import when scans with both pos and neg}

\item{peaksGroup}{\code{vector} to PeakGroupsParam(sampleGroups), should contain "QC"}

\item{centWaveParam}{xcms::CentWaveParam()}
}
\description{
Import \code{msDataFiles}, filter \code{ion_mode}, find peaks using \code{centWaveParam}, correct RT, group peaks using \code{peaksGroup}, fill peaks by xcms at MS1 Level
}
