% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSdev-function.R
\name{xcmsProcessingMSdev.DDA}
\alias{xcmsProcessingMSdev.DDA}
\title{xcmsProcessingMSdev.DDA}
\usage{
xcmsProcessingMSdev.DDA(object, ...)
}
\arguments{
\item{object}{MSdev}

\item{...}{}
}
\value{
MSdev
}
\description{
xcmsProcessingMSdev.DDA
}
