% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_openxlsx.R
\name{xlsx.write.list}
\alias{xlsx.write.list}
\title{xlsx.write.list}
\usage{
xlsx.write.list(df.list, file)
}
\arguments{
\item{df.list}{list of data}

\item{file}{file path}
}
\value{
null
}
\description{
Write a list of data.frame (or other object could be write into wroksheet), every term will be writen to a single sheet
}
