% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatisticFunction.R
\name{analyzePathwayHyperTest}
\alias{analyzePathwayHyperTest}
\title{analyzePathwayHypertest}
\usage{
analyzePathwayHyperTest(kegg.id)
}
\arguments{
\item{kegg.id}{}
}
\description{
analyzePathwayHypertest
}
