% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_inhouse_database.R,
%   R/build_inhouse_database_with_concentration.R
\name{build_inhouse_database}
\alias{build_inhouse_database}
\title{build inhouse database}
\usage{
build_inhouse_database(compound.record.file)

build_inhouse_database(compound.record.file)
}
\arguments{
\item{compound.record.file}{}
}
\description{
build inhouse database

build inhouse database
}
