% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_inhouse_database.R
\name{build_inhouse_database}
\alias{build_inhouse_database}
\title{build inhouse database}
\usage{
build_inhouse_database(compound.record.file)
}
\arguments{
\item{mzML.files}{}
}
\description{
build inhouse database
}
