% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_CFM.R
\name{CFM_data_get_igraph}
\alias{CFM_data_get_igraph}
\alias{cfm_data_get_fragment_group}
\title{CFM_data_get_igraph}
\usage{
CFM_data_get_igraph(object)

cfm_data_get_fragment_group(cfm_data, ppm = 10)
}
\arguments{
\item{object}{cfmd}

\item{cfm_data}{cfmd}

\item{ppm}{10}
}
\value{
cfmd

cfmd
}
\description{
CFM_data_get_igraph

cfm_data_get_fragment_group
}
\section{Functions}{
\itemize{
\item \code{CFM_data_get_igraph()}: CFM_data_get_igraph

\item \code{cfm_data_get_fragment_group()}: cfm_data_get_fragment_group

}}
