% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_CFM.R
\name{CFM_predict}
\alias{CFM_predict}
\title{CFM predict}
\usage{
CFM_predict(
  smiles_or_inchi_or_file = "[H]C1(O)O[C@]([H])(CO)[C@@]([H])(O)[C@]([H])(O)[C@@]1([H])O",
  prob_thresh = 0.001,
  param_adduct = "[M+H]+",
  annotate_fragments = 1,
  output_file_or_dir = NULL,
  apply_postproc = 0,
  suppress_exceptions = 1
)
}
\arguments{
\item{smiles_or_inchi_or_file}{smiles}

\item{prob_thresh}{prob}

\item{param_adduct}{adduct}

\item{annotate_fragments}{logic}

\item{output_file_or_dir}{path}

\item{apply_postproc}{logic}

\item{suppress_exceptions}{logic}
}
\value{
list of cfm data
}
\description{
see CFM doc
}
