% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_Spectra.R
\name{combineSpectra_groupby_ce}
\alias{combineSpectra_groupby_ce}
\title{combineSpectra_groupby_ce}
\usage{
combineSpectra_groupby_ce(sp, minProp = 0.5, ppm = 10, plot = F, ...)
}
\arguments{
\item{sp}{Spectra}

\item{minProp}{num}
}
\value{
Spectra
}
\description{
using \link{combinePeaksData_tic_weighted}, refer to \link[Spectra:combinePeaksData]{Spectra::combinePeaksData}
}
