% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_DEP.R
\name{DEP_export_data}
\alias{DEP_export_data}
\title{DEP_export_data
wirte coldata and rowdata to excel}
\usage{
DEP_export_data(data.se, file_path)
}
\arguments{
\item{data.se}{SE}

\item{file_path}{path}
}
\value{
null
}
\description{
DEP_export_data
wirte coldata and rowdata to excel
}
