% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_Chemmine.R
\name{get_sdf_igraph}
\alias{get_sdf_igraph}
\title{get_sdf_igraph}
\usage{
get_sdf_igraph(sdf, addH = F)
}
\arguments{
\item{sdf}{sdf}

\item{addH}{T or F}
}
\value{
igraph
}
\description{
get_sdf_igraph
}
