% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_xcms.R
\name{get_xcms_feature_ms1_candidate}
\alias{get_xcms_feature_ms1_candidate}
\title{match feature to database based on mz and rt
MSDB_id in db with mz error < mz.ppm will b}
\usage{
get_xcms_feature_ms1_candidate(xcms.xcms, db.path, mz.ppm = 25, rt.tol = Inf)
}
\arguments{
\item{rt.tol}{}
}
\description{
match feature to database based on mz and rt
MSDB_id in db with mz error < mz.ppm will b
}
