% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_xcms.R
\name{get_xcms_peaks_chrom}
\alias{get_xcms_peaks_chrom}
\title{get_xcms_peaks_chrom}
\usage{
get_xcms_peaks_chrom(xcms.xcms, peaks.id, all.sample = F, rt.range = "expand")
}
\arguments{
\item{rt}{one of c("all","identity","expand")}
}
\value{
XChromatograms
}
\description{
extract chromatograph from XCMSnExp,
if \code{all.sample} = F, only the samples, in which given peaks.id are detected will be return,
else extract from all samples
}
