% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSdev-function.R
\name{getStaDataMSdev}
\alias{getStaDataMSdev}
\title{getStaDataMSdev}
\usage{
getStaDataMSdev(
  object,
  MSDB.keys = c("Compound_name", "adduct", "formula", "inchikey", "Lipid_subclass",
    "database_origin")
)
}
\arguments{
\item{MSDB.keys}{}
}
\description{
getStaDataMSdev
}
