% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_adduct_to_peaks.R
\name{match_adduct_to_peaks}
\alias{match_adduct_to_peaks}
\title{match_adduct_to_peaks}
\usage{
match_adduct_to_peaks(MS.network, xcms.xcms, ppm.thresh = 10, rt.tol = 10)
}
\arguments{
\item{ppm.thresh}{}
}
\description{
match adduct candidate in xcms.peaks table
}
