% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_math.R
\name{match_mz}
\alias{match_mz}
\title{match 2 list of ion based on mz
return the closest match}
\usage{
match_mz(mz1, mz2, mz.ppm = 10)
}
\arguments{
\item{mz1}{num}

\item{mz2}{num}

\item{mz.ppm}{num}
}
\value{
num
}
\description{
match 2 list of ion based on mz
return the closest match
}
