% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_math.R
\name{match_mz_rt}
\alias{match_mz_rt}
\title{match 2 list of ion based on mz and rt}
\usage{
match_mz_rt(
  mz1,
  rt1 = seq_along(mz1),
  mz2,
  rt2 = seq_along(mz2),
  mz.ppm = 10,
  rt.tol = Inf
)
}
\arguments{
\item{rt.tol}{}
}
\description{
match 2 list of ion based on mz and rt
}
