% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metabolomic_workflow.R
\name{metabolomic_workflow}
\alias{metabolomic_workflow}
\title{Title}
\usage{
metabolomic_workflow(
  project.dir = "d:/2022_07_05-Lirui/",
  raw.data.dir = "d:/2022_07_05-Lirui/Data/"
)
}
\arguments{
\item{raw.data.dir}{}
}
\description{
Title
}
