% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSdev-function.R
\name{MSdev_annotation}
\alias{MSdev_annotation}
\title{MSdev annotation}
\usage{
MSdev_annotation(object, cpdb_path, ...)
}
\arguments{
\item{object}{MSdev}

\item{...}{}

\item{db.path}{CompoundDB}
}
\value{
MSdev
}
\description{
MSdev annotation
}
