% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSdev-function.R
\name{MSdev_extract_Spectra}
\alias{MSdev_extract_Spectra}
\title{MSdev_extract_Spectra}
\usage{
MSdev_extract_Spectra(object, msLevel = 2)
}
\arguments{
\item{object}{MSdev}
}
\value{
MSdev
}
\description{
MSdev_extract_Spectra
}
