% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSIP-function.R
\name{MSdev_find_isotope_label}
\alias{MSdev_find_isotope_label}
\title{MSdev_find_isotope_label}
\usage{
MSdev_find_isotope_label(object, iso_ele = "[13]C", ppm = 10, ...)
}
\arguments{
\item{object}{MSdev}

\item{ppm}{20}

\item{isotope}{"\link{13}C"}

\item{max_label}{10}

\item{net.degree.ratio}{0.5}
}
\value{
MSdev
}
\description{
MSdev_find_isotope_label
}
