% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_xcms.R
\name{plot_xcms_peaks_Chromatogram}
\alias{plot_xcms_peaks_Chromatogram}
\title{plot_xcms_peaks_Chromatogram}
\usage{
plot_xcms_peaks_Chromatogram(xcms.xcms, peak_id, rt = "expand")
}
\arguments{
\item{rt_expand}{foldchange to expand rt range}
}
\description{
extract EIC according to peaks' mzrange and rtrange,
note that if multiple sample in xcms object, only first sample will be extracted
}
