% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_xcms_peaks_distribution.R
\name{plot_xcms_peaks_distribution}
\alias{plot_xcms_peaks_distribution}
\title{plot_xcms_peaks_distribution}
\usage{
plot_xcms_peaks_distribution(xcms.xcms, plot.title = "Peaks distribution")
}
\arguments{
\item{plot.title}{}
}
\description{
export peaks data by xcms::chromPeaks and plot by ggplot2
}
