% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_CFM.R
\name{read_CFM_predict_result}
\alias{read_CFM_predict_result}
\title{read_CFM_predict_result}
\usage{
read_CFM_predict_result(result_path)
}
\arguments{
\item{result_path}{path}
}
\value{
list of cfm data
}
\description{
read_CFM_predict_result
}
