% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_Chemmine.R
\name{vis_sdf_igraph}
\alias{vis_sdf_igraph}
\title{vis_sdf_igraph}
\usage{
vis_sdf_igraph(
  sdf.igraph,
  show.id = F,
  prob.border = NULL,
  prob.fill = NULL,
  highlight = NULL
)
}
\arguments{
\item{sdf.igraph}{igraph}

\item{show.id}{logic}

\item{highlight}{vector}
}
\value{
vis html
}
\description{
vis_sdf_igraph
}
