% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_xcms.R
\name{xcmsProcessingMS1}
\alias{xcmsProcessingMS1}
\title{xcmsProcessingMS1}
\usage{
xcmsProcessingMS1(xcms.xcms, ion_mode = NA, xcms_param = NULL, ...)
}
\arguments{
\item{ion_mode}{to filter ion_mode, 1: positive, 0: negative, import when scans with both pos and neg}

\item{msDataFiles}{\code{char} ms file (full) paths}

\item{peaksGroup}{\code{vector} to PeakGroupsParam(sampleGroups), should contain "QC"}

\item{centWaveParam}{xcms::CentWaveParam()}
}
\value{
xcms
}
\description{
Import \code{msDataFiles}, filter \code{ion_mode}, find peaks using \code{centWaveParam}, correct RT, group peaks using \code{peaksGroup}, fill peaks by xcms at MS1 Level
}
